add test for battery
