+++ tesla_api/battery.py
@@ -3,12 +3,16 @@ from .base import Stub
 
 class Battery(Stub):
     @property
+    def not_enough_power_to_heat(self):
+        return self._vehicle._data["charge_state"]["not_enough_power_to_heat"]
+
+    @property
     def heater_on(self):
-        return self._vehicle.data["charge_state"]["battery_heater_on"]
+        return self._vehicle._data["charge_state"]["battery_heater_on"]
 
     @property
     def level(self) -> int:
-        return self._vehicle.data["charge_state"]["battery_level"]
+        return self._vehicle._data["charge_state"]["battery_level"]
 
     @property
     def soc(self) -> int:
@@ -16,19 +20,19 @@ class Battery(Stub):
 
     @property
     def usable_level(self) -> int:
-        return self._vehicle.data["charge_state"]["usable_battery_level"]
+        return self._vehicle._data["charge_state"]["usable_battery_level"]
 
     @property
     def range(self) -> int:
         # What format
-        return self._vehicle.data["charge_state"]["battery_range"]
+        return self._vehicle._data["charge_state"]["battery_range"]
 
     @property
     def estimated_range(self) -> int:
         # What format
-        return self._vehicle.data["charge_state"]["est_battery_range"]
+        return self._vehicle._data["charge_state"]["est_battery_range"]
 
     @property
     def ideal_range(self) -> int:
         # What format
-        return self._vehicle.data["charge_state"]["ideal_battery_range"]
+        return self._vehicle._data["charge_state"]["ideal_battery_range"]
